+++ changelog.md
@@ -2,6 +2,9 @@
 
 * Fixed forge server
 * Villager no longer make surprise sounds while trading
+* Fixed staying and following commands causing high CPU usage
+    * Panicking staying villagers will now run
+        * They will not return to original point yet, will be fixed in guards-update
 
 # 7.3.7